move the direct debit query to its own package
