+++ payment-service/src/main/java/com/hedvig/paymentservice/query/directDebit/DirectDebitAccountOrderRepository.kt
@@ -1,4 +1,4 @@
-package com.hedvig.paymentservice.query.member.entities
+package com.hedvig.paymentservice.query.directDebit
 
 import org.springframework.data.repository.CrudRepository
 import org.springframework.stereotype.Repository
